requirements.txt
torch
torchvision
torchaudio
torch-geometric
scikit-learn
networkx
javalang
tqdm
requests